Huge commit, not fully working but almost ready.
This is a huge commit with the goal of backing up the work done so far.
It probably won't work and deploy because all posts are draft and that
poses some problem when trying to create the static website (this bug
should be solved, but I don't expect it to happen once the website is
up).