Added ImageWithCaption component.
The ImageWithCaption component is used to add caption and framing on
images inserted with the markdown syntax. Some use cases, however, are
still uncovered by this component. For example, when the caption has
a different styled note (e.g., image attribution), the ImageWithCaption
cannot be used. Furthermore, the ImageWithCaption component uses alt
to fill the caption, which means that we can't use markdown or HTML in
it, as it will not be rendered.